Add files via upload
The 'pair_to_DB()' function now takes an additional parameter task_description which is the text entered by the contributor. This text is stored as a field in the outer document along with the pair_id, training set, and test set documents.

The function 'sol_to_DB()' now retrieves the task_description field from the Firestore document corresponding to the file, and includes it in the sol_document dictionary that is returned by the function.